+++ tie_break/src/pileup.rs
@@ -1,4 +1,3 @@
 
 pub mod csv_reader;
 pub mod shared_iterator;
-pub mod bam_reader;